Add DW alignment check for WIMM
There's a limitation in the H/W that for a WIMM operation, if the
pattern is a 8-byte one, the destination address must be 8-byte aligned.

The WIMM API in the Linux DMA framework always do with 8-byte, add a check of
alignment in switchtec_dma_prep_desc and return NULL if alignment requirement
not met.