Remove the workaround for the firmware fabric event issue
When the fabric event SWITCHTEC_FABRIC_REGISTER_BUFFER comes, firmware
should be able to accept the peer buffer retrieving initiated by API
switchtec_fabric_get_peer_buffers(). Firmware used to have issue
handling this retrieving properly, which is the reason we have this
delay to work around this.

Remove this workaround as firmware has fixed it.